requirements.txt
aniso8601==9.0.1
autopep8==1.5.6
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2==2.8.6
pycodestyle==2.7.0
python-dotenv==0.16.0
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.4
toml==0.10.2
Werkzeug==1.0.1
